pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "", email = ""},
]
dependencies = [
    "ipywidgets>=8.0.4",
    "matplotlib>=3.6.3",
    "pandas>=1.5.3",
    "scikit-learn>=1.1.2",
    "scipy>=1.9.3",
    "tqdm>=4.64.1",
    "optuna>=3.1.0",
    "lightgbm>=3.3.3",
    "py-ecg-detectors>=1.3.2",
    "PyWavelets>=1.4.1",
    "xarray>=2023.1.0",
    "statsmodels>=0.13.5",
    "petastorm>=0.12.1",
    "numba>=0.56.4",
    "skfeature-chappers>=1.1.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.12.0",
    "pre-commit>=3.0.2",
]

[tool.pdm.scripts]
post_install = "pre-commit install"
